requirements/local.txt
backports.shutil-get-terminal-size==1.0.0
coreapi==1.32.0
decorator==4.0.10
Django==1.9.9
django-filter==0.13.0
django-model-utils==2.5
django-rest-swagger==2.0.1
djangorestframework==3.4.0
honcho==0.7.1
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
itypes==1.1.0
Markdown==2.6.6
openapi-codec==1.0.0
path.py==8.2.1
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
PyYAML==3.11
requests==2.10.0
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.2.2
Unipath==1.1
uritemplate==0.6
wcwidth==0.1.7
